Restructuring: Dir & Files related to Goals feature
