+++ src/pages/AddGoalsPage/AddGoalsPage.tsx
@@ -5,14 +5,14 @@ import { useRecoilValue } from "recoil";
 import { useLocation } from "react-router";
 
 import { darkModeState } from "@store";
-import { GoalsForm } from "@components/GoalsComponents/GoalsForm";
+import { GoalsForm } from "@components/GoalsComponents/GoalsForm/GoalsForm";
 import { HeaderDashboard } from "@components/HeaderDashboard/HeaderDashboard";
 import { getGoal } from "@src/api/GoalsAPI";
 import { colorPallete } from "@src/utils";
 import paintBrush from "@assets/images/paintBrush.svg";
 
 import "@translations/i18n";
-import "@components/GoalsComponents/GoalsComponents.scss";
+import "@components/GoalsComponents/GoalsForm/GoalsComponents.scss";
 
 export const AddGoalsPage = () => {
   const darkModeStatus = useRecoilValue(darkModeState);
